+++ app/frontend/src/pages/MogacoPost/Controller/PostGroupId.tsx
@@ -1,0 +1,29 @@
+import { Controller, Control } from 'react-hook-form';
+
+import { Input } from '@/components';
+import { MOGACO_POST } from '@/constants';
+import { MogacoPostForm } from '@/types';
+
+type PostGroupIdProps = {
+  control: Control<MogacoPostForm>;
+};
+
+export function PostGroupId({ control }: PostGroupIdProps) {
+  return (
+    <Controller
+      control={control}
+      name="groupId"
+      rules={{ required: true }}
+      render={({ field: { onChange, value }, fieldState: { error } }) => (
+        <Input
+          label={MOGACO_POST.GROUP.LABEL}
+          placeholder={MOGACO_POST.GROUP.REQUIRED}
+          required
+          onChange={onChange}
+          value={value}
+          errorMessage={error && MOGACO_POST.GROUP.REQUIRED}
+        />
+      )}
+    />
+  );
+}
