+++ app/frontend/src/pages/MogacoPost/Controller/index.ts
@@ -1,0 +1,2 @@
+export * from './PostMemberId';
+export * from './PostTitle';
