+++ app/frontend/src/pages/MogacoPost/index.tsx
@@ -1,33 +1,23 @@
-import { useForm, Controller } from 'react-hook-form';
+import { useForm } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
 
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import dayjs from 'dayjs';
+import { Button } from '@/components';
+import { useSubmitPost } from '@/queries/hooks/post';
+import { MogacoPostForm } from '@/types';
 
-import { Input, Button, Textarea } from '@/components';
-import { MOGACO_POST } from '@/constants';
-import { queryKeys } from '@/queries';
-import { mogaco } from '@/services';
-import { MogacoPostForm, MogacoPostRequest } from '@/types';
-
+import { PostMemberId, PostTitle } from './Controller';
+import { PostAddress } from './Controller/PostAddress';
+import { PostContents } from './Controller/PostContents';
+import { PostDate } from './Controller/PostDate';
+import { PostGroupId } from './Controller/PostGroupId';
+import { PostMaxHumanCount } from './Controller/PostMaxHumanCount';
 import * as styles from './index.css';
-import { MogacoPostTitle } from './MogacoPostTitle';
 
 export function MogacoPostPage() {
-  const currentDate = dayjs().format('YYYY-MM-DD HH:mm');
   const { control, handleSubmit } = useForm<MogacoPostForm>();
 
   const navigate = useNavigate();
-
-  const queryClient = useQueryClient();
-  const { mutateAsync } = useMutation({
-    mutationFn: (form: MogacoPostRequest) => mogaco.post(form),
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: queryKeys.mogaco.list().queryKey,
-      });
-    },
-  });
+  const { mutateAsync } = useSubmitPost();
 
   const onSubmit = async ({
     title,
@@ -54,124 +44,14 @@ export function MogacoPostPage() {
 
   return (
     <form className={styles.container} onSubmit={handleSubmit(onSubmit)}>
-      <MogacoPostTitle control={control} />
+      <PostTitle control={control} />
       <div className={styles.formContent}>
-        <Controller
-          control={control}
-          name="memberId"
-          render={({ field: { onChange, value } }) => (
-            <Input
-              label={MOGACO_POST.MEMBER.LABEL}
-              required
-              disabled
-              defaultValue="user" // 로그인한 유저 정보
-              onChange={onChange}
-              value={value}
-            />
-          )}
-        />
-        <Controller
-          control={control}
-          name="groupId"
-          rules={{ required: true }}
-          render={({ field: { onChange, value }, fieldState: { error } }) => (
-            <Input
-              label={MOGACO_POST.GROUP.LABEL}
-              placeholder={MOGACO_POST.GROUP.REQUIRED}
-              required
-              onChange={onChange}
-              value={value}
-              errorMessage={error && MOGACO_POST.GROUP.REQUIRED}
-            />
-          )}
-        />
-        <Controller
-          control={control}
-          name="maxHumanCount"
-          rules={{
-            required: MOGACO_POST.COUNT.REQUIRED,
-            pattern: {
-              value: /^[0-9]*$/,
-              message: MOGACO_POST.COUNT.PATTERN,
-            },
-            min: {
-              value: MOGACO_POST.COUNT.MIN_VALUE,
-              message: MOGACO_POST.COUNT.MIN,
-            },
-            max: {
-              value: MOGACO_POST.COUNT.MAX_VALUE,
-              message: MOGACO_POST.COUNT.MAX,
-            },
-          }}
-          render={({ field: { onChange, value }, fieldState: { error } }) => (
-            <Input
-              label={MOGACO_POST.COUNT.LABEL}
-              type="number"
-              placeholder={MOGACO_POST.COUNT.REQUIRED}
-              required
-              onChange={onChange}
-              value={value}
-              errorMessage={error && error.message}
-            />
-          )}
-        />
-        <Controller
-          control={control}
-          name="address"
-          rules={{ required: true }}
-          render={({ field: { onChange, value }, fieldState: { error } }) => (
-            <Input
-              label={MOGACO_POST.ADDRESS.LABEL}
-              placeholder={MOGACO_POST.ADDRESS.REQUIRED}
-              required
-              onChange={onChange}
-              value={value}
-              errorMessage={error && MOGACO_POST.ADDRESS.REQUIRED}
-            />
-          )}
-        />
-        <Controller
-          control={control}
-          name="date"
-          rules={{
-            required: MOGACO_POST.DATE.REQUIRED,
-            min: { value: currentDate, message: MOGACO_POST.DATE.MIN },
-          }}
-          render={({ field: { onChange, value }, fieldState: { error } }) => (
-            <Input
-              label={MOGACO_POST.DATE.LABEL}
-              type="datetime-local"
-              required
-              min={currentDate}
-              onChange={onChange}
-              value={value}
-              errorMessage={error && error.message}
-            />
-          )}
-        />
-        <Controller
-          control={control}
-          name="contents"
-          rules={{
-            required: MOGACO_POST.CONTENTS.REQUIRED,
-            maxLength: {
-              value: MOGACO_POST.CONTENTS.MAX_LENGTH,
-              message: MOGACO_POST.CONTENTS.MAX,
-            },
-          }}
-          render={({ field: { onChange, value }, fieldState: { error } }) => (
-            <Textarea
-              label={MOGACO_POST.CONTENTS.LABEL}
-              placeholder={MOGACO_POST.CONTENTS.REQUIRED}
-              rows={MOGACO_POST.CONTENTS.ROWS}
-              maxLength={MOGACO_POST.CONTENTS.MAX_LENGTH}
-              required
-              onChange={onChange}
-              value={value}
-              errorMessage={error && error.message}
-            />
-          )}
-        />
+        <PostMemberId control={control} />
+        <PostGroupId control={control} />
+        <PostMaxHumanCount control={control} />
+        <PostAddress control={control} />
+        <PostDate control={control} />
+        <PostContents control={control} />
       </div>
       <div className={styles.formContent}>
         <Button
