+++ app/frontend/src/queries/hooks/post.ts
@@ -1,0 +1,18 @@
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+
+import { queryKeys } from '@/queries';
+import { mogaco } from '@/services';
+import { MogacoPostRequest } from '@/types';
+
+export const useSubmitPost = () => {
+  const queryClient = useQueryClient();
+
+  return useMutation({
+    mutationFn: (form: MogacoPostRequest) => mogaco.post(form),
+    onSuccess: () => {
+      queryClient.invalidateQueries({
+        queryKey: queryKeys.mogaco.list().queryKey,
+      });
+    },
+  });
+};
